% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompareCohorts.R
\name{computeStandardizedDifference}
\alias{computeStandardizedDifference}
\title{Compute standardized difference of mean for all covariates.}
\usage{
computeStandardizedDifference(covariateData1, covariateData2)
}
\arguments{
\item{covariateData1}{The covariate data of the first cohort. Needs to be in aggregated format.}

\item{covariateData2}{The covariate data of the second cohort. Needs to be in aggregated format.}
}
\value{
A data frame with means and standard deviations per cohort as well as the standardized 
difference of mean.
}
\description{
Computes the standardized difference for all covariates between two cohorts. The standardized
difference is defined as the difference between the mean divided by the overall standard deviation.
}
