% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetailedCovariateSettings.R
\name{convertPrespecSettingsToDetailedSettings}
\alias{convertPrespecSettingsToDetailedSettings}
\title{Convert prespecified covariate settings into detailed covariate settings}
\usage{
convertPrespecSettingsToDetailedSettings(covariateSettings)
}
\arguments{
\item{covariateSettings}{An object of type \code{covariateSettings} as created for example by the
\code{\link{createCovariateSettings}} function.}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Convert prespecified covariate settings into detailed covariate settings
}
\details{
For advanced users only.
}
\examples{
\donttest{
covSettings <- createDefaultCovariateSettings()
detailedSettings <- convertPrespecSettingsToDetailedSettings(covariateSettings = covSettings)
}

}
