% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateData.R
\docType{class}
\name{CovariateData-class}
\alias{CovariateData-class}
\alias{show,CovariateData-method}
\alias{summary,CovariateData-method}
\title{CovariateData class.}
\usage{
\S4method{show}{CovariateData}(object)

\S4method{summary}{CovariateData}(object)
}
\description{
CovariateData class.
}
\keyword{internal}
