% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CareSiteSettings.R
\name{createCareSiteEraCovariateSettings}
\alias{createCareSiteEraCovariateSettings}
\title{Create care-site covariate settings from the care_site_era table}
\usage{
createCareSiteEraCovariateSettings(
  analysisId,
  startDays = -365,
  endDays = 0,
  startReferenceDate = "cohort_start_date",
  endReferenceDate = "cohort_start_date",
  includeAncestorConcepts = TRUE,
  includedCareSiteClassIds = NULL
)
}
\arguments{
\item{analysisId}{Unique integer between 0 and 999}

\item{startDays}{Integer days relative to startReferenceDate or "anyTimePrior"}

\item{endDays}{Integer days relative to endReferenceDate or "anyTimeAfter"}

\item{startReferenceDate}{Construct time-interval for covariate starting relative to 
"cohort_start_date" or "cohort_end_date"}

\item{endReferenceDate}{Construct time-interval for covariate ending relative toBibtex
"cohort_start_date" or "cohort_end_date"}

\item{includeAncestorConcepts}{Boolean: include ancestor concepts to identified care_site_concept_id}

\item{includedCareSiteClassIds}{Restrict included concepts to care_site_class_id in given vector}
}
\value{
An object of type \code{covariateSettings} to be used in other functions.
}
\description{
Create care-site covariate settings from the care_site_era table
}
\details{
Creates an object specifying how care-site covariates should be constructed from 
care_site_era in the CDM model
}
