% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCovariatesSparse.R
\name{createCovariateSettingsSparse}
\alias{createCovariateSettingsSparse}
\title{Create covariate settings}
\usage{
createCovariateSettingsSparse(plpData, studyPop, dataSplit,
  numBasisVectors = -1, useUsualCovariatesAlso = TRUE, lambda1 = 0.2,
  numThreads = -1, batchsize = 400, iter = 1000, posAlpha = FALSE,
  posD = TRUE, eps = 0.001)
}
\arguments{
\item{plpData}{A plpData object}

\item{studyPop}{}

\item{dataSplit}{}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how new features should be contructed through sparse coding from
standardly encoded covariates.
}
