% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetailedCovariateSettings.R
\name{createDefaultCovariateSettings}
\alias{createDefaultCovariateSettings}
\title{Create default covariate settings}
\usage{
createDefaultCovariateSettings()
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create default covariate settings
}
