% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateData.R
\name{createEmptyCovariateData}
\alias{createEmptyCovariateData}
\title{Create an empty CovariateData object}
\usage{
createEmptyCovariateData(cohortId, cohortIds, aggregated, temporal)
}
\arguments{
\item{cohortId}{DEPRECATED}

\item{cohortIds}{For which cohort ID(s) should covariates be constructed? If set to c(-1),
covariates will be constructed for all cohorts in the specified cohort
table.}

\item{aggregated}{if the covariate data is aggregated}

\item{temporal}{if the covariate data is temporal}
}
\value{
A logical value.
}
\description{
Create an empty CovariateData object
}
