% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CareSiteSettings.R
\name{createPersonCareSiteCovariateSettings}
\alias{createPersonCareSiteCovariateSettings}
\title{Create care-site covariate settings from the person table}
\usage{
createPersonCareSiteCovariateSettings(
  analysisId,
  includeAncestorConcepts = TRUE,
  includedCareSiteClassIds = NULL
)
}
\arguments{
\item{analysisId}{Unique integer between 0 and 999}

\item{includeAncestorConcepts}{Boolean: include ancestor concepts to identified care_site_concept_id}

\item{includedCareSiteClassIds}{Restrict included concepts to care_site_class_id in given vector}
}
\value{
An object of type \code{covariateSettings} to be used in other functions.
}
\description{
Create care-site covariate settings from the person table
}
\details{
Creates an object specifying how care-site covariates should be constructed from 
person.care_site_id in the CDM model
}
