% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnitTestHelperFunctions.R
\name{.createLooCovariateSettings}
\alias{.createLooCovariateSettings}
\title{Get covariate settings}
\usage{
.createLooCovariateSettings(useLengthOfObs = TRUE)
}
\arguments{
\item{useLengthOfObs}{if length of observations should be used}
}
\value{
Returns an object of type \code{covariateSettings}, containing settings for the covariates.
}
\description{
Get covariate settings
}
