% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction.R
\docType{package}
\name{FeatureExtraction-package}
\alias{FeatureExtraction}
\alias{FeatureExtraction-package}
\title{FeatureExtraction: Generating Features for a Cohort}
\description{
An R interface for generating features for a cohort using data in the Common Data Model. Features can be constructed using default or custom made feature definitions. Furthermore it's possible to aggregate features and get the summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/FeatureExtraction}
  \item Report bugs at \url{https://github.com/OHDSI/FeatureExtraction/issues}
}

}
\author{
\strong{Maintainer}: Ger Inberg \email{g.inberg@erasmusmc.nl}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Marc Suchard
  \item Patrick Ryan
  \item Jenna Reps
  \item Anthony Sena \email{sena@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
