% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterByCohortDefinitionId}
\alias{filterByCohortDefinitionId}
\title{Filter covariates by cohort definition IDs}
\usage{
filterByCohortDefinitionId(covariateData, cohortIds = c(1))
}
\arguments{
\item{covariateData}{An object of type \code{CovariateData}}

\item{cohortIds}{The cohort definition IDs to keep.}
}
\value{
An object of type \code{covariateData}.
}
\description{
Filter covariates by cohort definition IDs
}
\examples{
\dontrun{
covariateData <- FeatureExtraction:::createEmptyCovariateData(
  cohortIds = 1,
  aggregated = FALSE,
  temporal = FALSE
)

covData <- filterByCohortDefinitionId(
  covariateData = covariateData,
  cohortIds = c(1)
)
}

}
