% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterByRowId}
\alias{filterByRowId}
\title{Filter covariates by row ID}
\usage{
filterByRowId(object, rowIds)
}
\arguments{
\item{object}{Either an object of type \code{covariateData}, or an ffdf object containing
covariate values.}

\item{rowIds}{A vector (or ff object) containing the rowIds to keep.}
}
\value{
Either an object of type \code{covariateData}, or an ffdf object containing
                covariate values. (depending on the type of the \code{object} argument.
}
\description{
Filter covariates by row ID
}
