% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefaultCovariates.R
\name{filterCovariateDataCovariates}
\alias{filterCovariateDataCovariates}
\title{Filters the covariateData covariates based on the given characterization mean value.}
\usage{
filterCovariateDataCovariates(
  covariateData,
  covariatesName,
  minCharacterizationMean = 0
)
}
\arguments{
\item{covariateData}{The covariate data}

\item{covariatesName}{The name of the covariates object inside the covariateData}

\item{minCharacterizationMean}{The minimum mean value for characterization output. Values below this will be cut off from output. This
will help reduce the file size of the characterization output, but will remove information
on covariates that have very low values. The default is 0.}
}
\description{
Filters the covariateData covariates based on the given characterization mean value.
}
