% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table1.R
\name{getDefaultTable1Specifications}
\alias{getDefaultTable1Specifications}
\title{Get the default table 1 specifications}
\usage{
getDefaultTable1Specifications()
}
\value{
A specifications objects.
}
\description{
Loads the default specifications for a table 1, to be used with the \code{\link{createTable1}}
function.
}
\examples{
\donttest{
defaultTable1Specs <- getDefaultTable1Specifications()
}

}
