% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateData.R
\name{isAggregatedCovariateData}
\alias{isAggregatedCovariateData}
\title{Check whether covariate data is aggregated}
\usage{
isAggregatedCovariateData(x)
}
\arguments{
\item{x}{The covariate data object to check.}
}
\value{
A logical value.
}
\description{
Check whether covariate data is aggregated
}
