% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCovariatesFromOtherCohorts.R
\name{replaceCovariateSettingsCohortSchemaTable}
\alias{replaceCovariateSettingsCohortSchemaTable}
\title{Utility function to set the cohort table & schema on createCohortBasedCovariateSettings
with information from the execution settings}
\usage{
replaceCovariateSettingsCohortSchemaTable(
  covariateSettings,
  covariateCohortDatabaseSchema,
  covariateCohortTable
)
}
\arguments{
\item{covariateSettings}{An object of type \code{covariateSettings}}

\item{covariateCohortDatabaseSchema}{The database schema where the cohorts used to define the covariates can be found.}

\item{covariateCohortTable}{The table where the cohorts used to define the covariates can be found.}
}
\value{
An object of type \code{covariateSettings}
}
\description{
Utility function to set the cohort table & schema on createCohortBasedCovariateSettings
with information from the execution settings
}
