% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{tidyCovariateData}
\alias{tidyCovariateData}
\title{Tidy covariate data}
\usage{
tidyCovariateData(covariateData, normalize = TRUE, removeRedundancy = TRUE)
}
\arguments{
\item{covariateData}{An object as generated using the \code{\link{getDbCovariateData}}
function.}

\item{normalize}{Normalize the coviariates? (dividing by the max)}

\item{removeRedundancy}{Should redundant covariates be removed?}
}
\description{
Tidy covariate data
}
\details{
Normalize covariate values by dividing by the max and/or remove redundant covariates.
}
