% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aggregation.R
\name{aggregateCovariates}
\alias{aggregateCovariates}
\title{Aggregate covariate data}
\usage{
aggregateCovariates(covariateData)
}
\arguments{
\item{covariateData}{An object of type \code{covariateData} as generated using
\code{getDbCovariateData}.}
}
\value{
An object of class \code{covariateData}.
}
\description{
Aggregate covariate data
}
\examples{
\donttest{
covariateData <- FeatureExtraction::createEmptyCovariateData(
  cohortIds = 1,
  aggregated = FALSE,
  temporal = FALSE
)
aggregatedCovariateData <- aggregateCovariates(covariateData)
}

}
