% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCovariateSettings.R
\name{createCovariateSettings}
\alias{createCovariateSettings}
\title{Create covariate settings}
\usage{
createCovariateSettings(useDemographicsGender = FALSE,
  useDemographicsAge = FALSE, useDemographicsIndexYear = FALSE,
  useDemographicsIndexMonth = FALSE, useConditionOccurrenceLongTerm = FALSE,
  useConditionOccurrenceShortTerm = FALSE, useConditionEraLongTerm = FALSE,
  useConditionEraShortTerm = FALSE, useConditionGroupEraShortTerm = FALSE,
  useConditionGroupEraLongTerm = FALSE, useDrugExposureLongTerm = FALSE,
  useDrugExposureShortTerm = FALSE, useDrugEraLongTerm = FALSE,
  useDrugEraShortTerm = FALSE, useDrugGroupEraLongTerm = FALSE,
  useDrugGroupEraShortTerm = FALSE, useProcedureOccurrenceLongTerm = FALSE,
  useProcedureOccurrenceShortTerm = FALSE,
  useDeviceExposureLongTerm = FALSE, useDeviceExposureShortTerm = FALSE,
  useMeasurementLongTerm = FALSE, useMeasurementShortTerm = FALSE,
  useObservationLongTerm = FALSE, useObservationShortTerm = FALSE,
  useCharlsonIndex = FALSE, longTermDays = 365, shortTermDays = 30,
  windowEndDays = 0, excludedCovariateConceptIds = c(),
  addDescendantsToExclude = TRUE, includedCovariateConceptIds = c(),
  addDescendantsToInclude = TRUE, includedCovariateIds = c(),
  deleteCovariatesSmallCount = 100)
}
\arguments{
\item{useDemographicsGender}{Gender of the subject.}

\item{useDemographicsAge}{Age of the subject on the index date (in 5 year
increments).}

\item{useDemographicsIndexYear}{Year of the index date.}

\item{useDemographicsIndexMonth}{Month of the index date.}

\item{useConditionOccurrenceLongTerm}{One covariate per condition in the condition_occurrence
table starting in the long term window.}

\item{useConditionOccurrenceShortTerm}{One covariate per condition in the condition_occurrence
table starting in the short term window.}

\item{useConditionEraLongTerm}{One covariate per condition in the condition_era table
overlapping with any part of the long term window.}

\item{useConditionEraShortTerm}{One covariate per condition in the condition_era table
overlapping with any part of the short term window.}

\item{useConditionGroupEraShortTerm}{One covariate per condition era rolled up to SNOMED groups
in the condition_era table overlapping with any part of the
long term window.}

\item{useConditionGroupEraLongTerm}{One covariate per condition era rolled up to SNOMED groups
in the condition_era table overlapping with any part of the
short term window.}

\item{useDrugExposureLongTerm}{One covariate per drug in the drug_exposure table starting
in the long term window.}

\item{useDrugExposureShortTerm}{One covariate per drug in the drug_exposure table starting
in the short term window.}

\item{useDrugEraLongTerm}{One covariate per drug in the drug_era table overlapping
with any part of the long term window.}

\item{useDrugEraShortTerm}{One covariate per drug in the drug_era table overlapping
with any part of the long short window.}

\item{useDrugGroupEraLongTerm}{One covariate per drug rolled up to ATC groups in the
drug_era table overlapping with any part of the long term
window.}

\item{useDrugGroupEraShortTerm}{One covariate per drug rolled up to ATC groups in the
drug_era table overlapping with any part of the short term
window.}

\item{useProcedureOccurrenceLongTerm}{One covariate per procedure in the procedure_occurrence
table in the long term window.}

\item{useProcedureOccurrenceShortTerm}{One covariate per procedure in the procedure_occurrence
table in the short term window.}

\item{useDeviceExposureLongTerm}{One covariate per device in the device exposure table
starting in the long term window.}

\item{useDeviceExposureShortTerm}{One covariate per device in the device exposure table
starting in the short term window.}

\item{useMeasurementLongTerm}{One covariate per measurement in the measurement table in
the long term window.}

\item{useMeasurementShortTerm}{One covariate per measurement in the measurement table in
the short term window.}

\item{useObservationLongTerm}{One covariate per observation in the observation table in
the long term window.}

\item{useObservationShortTerm}{One covariate per observation in the observation table in
the short term window.}

\item{useCharlsonIndex}{The Charlson comorbidity index (Romano adaptation) using
all conditions prior to the window end.}

\item{longTermDays}{What is the length (in days) of the long-term window?}

\item{shortTermDays}{What is the length (in days) of the short-term window?}

\item{windowEndDays}{What is the last day of the window? 0 means the cohort
start date is the last date (included), 1 means the window
stops the day before the cohort start date, etc.}

\item{mediumTermDays}{What is the length (in days) of the medium-term window?}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}
\examples{
\dontrun{
# This will create the default set of covariates:
settings <- createCovariateSettings(useDemographicsGender = TRUE,
                                    useDemographicsAge = TRUE,
                                    useDemographicsIndexYear = TRUE,
                                    useDemographicsIndexMonth = TRUE,
                                    useConditionOccurrenceLongTerm = TRUE,
                                    useConditionOccurrenceShortTerm = TRUE,
                                    useConditionEraLongTerm = TRUE,
                                    useConditionEraShortTerm = TRUE,
                                    useConditionGroupEraShortTerm = TRUE,
                                    useConditionGroupEraLongTerm = TRUE,
                                    useDrugExposureLongTerm = TRUE,
                                    useDrugExposureShortTerm = TRUE,
                                    useDrugEraLongTerm = TRUE,
                                    useDrugEraShortTerm = TRUE,
                                    useDrugGroupEraLongTerm = TRUE,
                                    useDrugGroupEraShortTerm = TRUE,
                                    useProcedureOccurrenceLongTerm = TRUE,
                                    useProcedureOccurrenceShortTerm = TRUE,
                                    useDeviceExposureLongTerm = TRUE,
                                    useDeviceExposureShortTerm = TRUE,
                                    useMeasurementLongTerm = TRUE,
                                    useMeasurementShortTerm = TRUE,
                                    useObservationLongTerm = TRUE,
                                    useObservationShortTerm = TRUE,
                                    useCharlsonIndex = TRUE)

}

}
