% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefaultCovariates.R
\name{createCovariateSettings}
\alias{createCovariateSettings}
\title{Create covariate settings}
\usage{
createCovariateSettings(useCovariateDemographics = FALSE,
  useCovariateDemographicsGender = FALSE,
  useCovariateDemographicsRace = FALSE,
  useCovariateDemographicsEthnicity = FALSE,
  useCovariateDemographicsAge = FALSE, useCovariateDemographicsYear = FALSE,
  useCovariateDemographicsMonth = FALSE,
  useCovariateConditionOccurrence = FALSE,
  useCovariateConditionOccurrenceLongTerm = FALSE,
  useCovariateConditionOccurrenceShortTerm = FALSE,
  useCovariateConditionOccurrenceInptMediumTerm = FALSE,
  useCovariateConditionEra = FALSE, useCovariateConditionEraEver = FALSE,
  useCovariateConditionEraOverlap = FALSE,
  useCovariateConditionGroup = FALSE,
  useCovariateConditionGroupMeddra = FALSE,
  useCovariateConditionGroupSnomed = FALSE,
  useCovariateDrugExposure = FALSE,
  useCovariateDrugExposureLongTerm = FALSE,
  useCovariateDrugExposureShortTerm = FALSE, useCovariateDrugEra = FALSE,
  useCovariateDrugEraLongTerm = FALSE, useCovariateDrugEraShortTerm = FALSE,
  useCovariateDrugEraOverlap = FALSE, useCovariateDrugEraEver = FALSE,
  useCovariateDrugGroup = FALSE, useCovariateProcedureOccurrence = FALSE,
  useCovariateProcedureOccurrenceLongTerm = FALSE,
  useCovariateProcedureOccurrenceShortTerm = FALSE,
  useCovariateProcedureGroup = FALSE, useCovariateObservation = FALSE,
  useCovariateObservationLongTerm = FALSE,
  useCovariateObservationShortTerm = FALSE,
  useCovariateObservationCountLongTerm = FALSE,
  useCovariateMeasurement = FALSE, useCovariateMeasurementLongTerm = FALSE,
  useCovariateMeasurementShortTerm = FALSE,
  useCovariateMeasurementCountLongTerm = FALSE,
  useCovariateMeasurementBelow = FALSE,
  useCovariateMeasurementAbove = FALSE, useCovariateConceptCounts = FALSE,
  useCovariateRiskScores = FALSE, useCovariateRiskScoresCharlson = FALSE,
  useCovariateRiskScoresDCSI = FALSE, useCovariateRiskScoresCHADS2 = FALSE,
  useCovariateRiskScoresCHADS2VASc = FALSE,
  useCovariateInteractionYear = FALSE, useCovariateInteractionMonth = FALSE,
  excludedCovariateConceptIds = c(), addDescendantsToExclude = TRUE,
  includedCovariateConceptIds = c(), addDescendantsToInclude = TRUE,
  deleteCovariatesSmallCount = 100, longTermDays = 365,
  mediumTermDays = 180, shortTermDays = 30,
  useCovariateProcedureOccurrence365d, useCovariateConditionOccurrence365d,
  useCovariateDrugExposure365d, useCovariateMeasurementCount365d,
  useCovariateDrugEra365d, useCovariateObservation365d,
  useCovariateObservationCount365d, useCovariateMeasurement365d,
  useCovariateConditionOccurrenceInpt180d, useCovariateConditionOccurrence30d,
  useCovariateDrugExposure30d, useCovariateDrugEra30d,
  useCovariateMeasurement30d, useCovariateObservation30d,
  useCovariateProcedureOccurrence30d)
}
\arguments{
\item{useCovariateDemographics}{A boolean value (TRUE/FALSE) to determine if
demographic covariates (age in 5-yr increments,
gender, race, ethnicity, year of index date, month
of index date) will be created and included in
future models.}

\item{useCovariateDemographicsGender}{A boolean value (TRUE/FALSE) to determine if gender
should be included in the model.}

\item{useCovariateDemographicsRace}{A boolean value (TRUE/FALSE) to determine if race
should be included in the model.}

\item{useCovariateDemographicsEthnicity}{A boolean value (TRUE/FALSE) to determine if
ethnicity should be included in the model.}

\item{useCovariateDemographicsAge}{A boolean value (TRUE/FALSE) to determine if age
(in 5 year increments) should be included in the
model.}

\item{useCovariateDemographicsYear}{A boolean value (TRUE/FALSE) to determine if
calendar year should be included in the model.}

\item{useCovariateDemographicsMonth}{A boolean value (TRUE/FALSE) to determine if
calendar month should be included in the model.}

\item{useCovariateConditionOccurrence}{A boolean value (TRUE/FALSE) to determine if
covariates derived from CONDITION_OCCURRENCE table
will be created and included in future models.}

\item{useCovariateConditionOccurrenceLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition in the long term
window prior to or on cohort index date.  Only
applicable if useCovariateConditionOccurrence =
TRUE.}

\item{useCovariateConditionOccurrenceShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition in the short term
window prior to or on cohort index date.  Only
applicable if useCovariateConditionOccurrence =
TRUE.}

\item{useCovariateConditionOccurrenceInptMediumTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition within
inpatient type in medium term window prior to or on cohort
index date.  Only applicable if
useCovariateConditionOccurrence = TRUE.}

\item{useCovariateConditionEra}{A boolean value (TRUE/FALSE) to determine if
covariates derived from CONDITION_ERA table will be
created and included in future models.}

\item{useCovariateConditionEraEver}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition era anytime
prior to or on cohort index date.  Only applicable
if useCovariateConditionEra = TRUE.}

\item{useCovariateConditionEraOverlap}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition era that
overlaps the cohort index date.  Only applicable if
useCovariateConditionEra = TRUE.}

\item{useCovariateConditionGroup}{A boolean value (TRUE/FALSE) to determine if all
CONDITION_OCCURRENCE and CONDITION_ERA covariates
should be aggregated or rolled-up to higher-level
concepts based on vocabluary classification.}

\item{useCovariateConditionGroupMeddra}{A boolean value (TRUE/FALSE) to determine if all
CONDITION_OCCURRENCE and CONDITION_ERA covariates
should be aggregated or rolled-up to higher-level
concepts based on the MEDDRA classification.}

\item{useCovariateConditionGroupSnomed}{A boolean value (TRUE/FALSE) to determine if all
CONDITION_OCCURRENCE and CONDITION_ERA covariates
should be aggregated or rolled-up to higher-level
concepts based on the SNOMED classification.}

\item{useCovariateDrugExposure}{A boolean value (TRUE/FALSE) to determine if
covariates derived from DRUG_EXPOSURE table will be
created and included in future models.}

\item{useCovariateDrugExposureLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug in the long term window
prior to or on cohort index date.  Only applicable
if useCovariateDrugExposure = TRUE.}

\item{useCovariateDrugExposureShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug in the short term window
prior to or on cohort index date.  Only applicable
if useCovariateDrugExposure = TRUE.}

\item{useCovariateDrugEra}{A boolean value (TRUE/FALSE) to determine if
covariates derived from DRUG_ERA table will be
created and included in future models.}

\item{useCovariateDrugEraLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era in the long term
window prior to or on cohort index date.  Only
applicable if useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era in the short term window
prior to or on cohort index date.  Only applicable
if useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraOverlap}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era that overlaps
the cohort index date.  Only applicable if
useCovariateDrugEra = TRUE.}

\item{useCovariateDrugEraEver}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of drug era anytime prior
to or on cohort index date.  Only applicable if
useCovariateDrugEra = TRUE.}

\item{useCovariateDrugGroup}{A boolean value (TRUE/FALSE) to determine if all
DRUG_EXPOSURE and DRUG_ERA covariates should be
aggregated or rolled-up to higher-level concepts of
drug classes based on vocabluary classification.}

\item{useCovariateProcedureOccurrence}{A boolean value (TRUE/FALSE) to determine if
covariates derived from PROCEDURE_OCCURRENCE table
will be created and included in future models.}

\item{useCovariateProcedureOccurrenceLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of procedure in the long term
window prior to or on cohort index date.  Only
applicable if useCovariateProcedureOccurrence =
TRUE.}

\item{useCovariateProcedureOccurrenceShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of procedure in the short term
window prior to or on cohort index date.  Only
applicable if useCovariateProcedureOccurrence =
TRUE.}

\item{useCovariateProcedureGroup}{A boolean value (TRUE/FALSE) to determine if all
PROCEDURE_OCCURRENCE covariates should be
aggregated or rolled-up to higher-level concepts
based on vocabluary classification.}

\item{useCovariateObservation}{A boolean value (TRUE/FALSE) to determine if
covariates derived from OBSERVATION table will be
created and included in future models.}

\item{useCovariateObservationLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of observation in the long term
window prior to or on cohort index date.  Only
applicable if useCovariateObservation = TRUE.}

\item{useCovariateObservationShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of observation in the short term
window prior to or on cohort index date.  Only
applicable if useCovariateObservation = TRUE.}

\item{useCovariateObservationCountLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for the count of each observation concept in
LongTerm window prior to or on cohort index date.  Only
applicable if useCovariateObservation = TRUE.}

\item{useCovariateMeasurement}{A boolean value (TRUE/FALSE) to determine if
covariates derived from OBSERVATION table will be
created and included in future models.}

\item{useCovariateMeasurementLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement in the long term
window prior to or on cohort index date.  Only
applicable if useCovariateMeasurement = TRUE.}

\item{useCovariateMeasurementShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement in the short term
window prior to or on cohort index date.  Only
applicable if useCovariateMeasurement = TRUE.}

\item{useCovariateMeasurementCountLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for the count of each measurement concept in
LongTerm window prior to or on cohort index date.  Only
applicable if useCovariateMeasurement = TRUE.}

\item{useCovariateMeasurementBelow}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement with a
numeric value below normal range for latest value
within medium term window of cohort index.  Only applicable if
useCovariateMeasurement = TRUE (CDM v5+) or
useCovariateObservation = TRUE (CDM v4).}

\item{useCovariateMeasurementAbove}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of measurement with a
numeric value above normal range for latest value
within medium term window of cohort index.  Only applicable if
useCovariateMeasurement = TRUE (CDM v5+) or
useCovariateObservation = TRUE (CDM v4).}

\item{useCovariateConceptCounts}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
count the number of concepts that a person has
within each domain (CONDITION, DRUG, PROCEDURE,
OBSERVATION)}

\item{useCovariateRiskScores}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
calculate various Risk Scores, including Charlson,
DCSI.}

\item{useCovariateRiskScoresCharlson}{A boolean value (TRUE/FALSE) to determine if the
Charlson comorbidity index should be included in
the model.}

\item{useCovariateRiskScoresDCSI}{A boolean value (TRUE/FALSE) to determine if the
DCSI score should be included in the model.}

\item{useCovariateRiskScoresCHADS2}{A boolean value (TRUE/FALSE) to determine if the
CHADS2 score should be included in the model.}

\item{useCovariateRiskScoresCHADS2VASc}{A boolean value (TRUE/FALSE) to determine if the
CHADS2VASc score should be included in the model.}

\item{useCovariateInteractionYear}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
represent interaction terms between all other
covariates and the year of the cohort index date.}

\item{useCovariateInteractionMonth}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
represent interaction terms between all other
covariates and the month of the cohort index date.}

\item{excludedCovariateConceptIds}{A list of concept IDs that should NOT be used to
construct covariates.}

\item{addDescendantsToExclude}{Should descendant concept IDs be added to the list 
of concepts to exclude?}

\item{includedCovariateConceptIds}{A list of concept IDs that should be used to
construct covariates.}

\item{addDescendantsToInclude}{Should descendant concept IDs be added to the list 
of concepts to include?}

\item{deleteCovariatesSmallCount}{A numeric value used to remove covariates that
occur in both cohorts fewer than
deleteCovariateSmallCounts time.}

\item{longTermDays}{What is the length (in days) of the long-term window?}

\item{mediumTermDays}{What is the length (in days) of the medium-term window?}

\item{shortTermDays}{What is the length (in days) of the short-term window?}

\item{useCovariateProcedureOccurrence365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateConditionOccurrence365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateDrugExposure365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateMeasurementCount365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateDrugEra365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateObservation365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateObservationCount365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateMeasurement365d}{DEPRECATED. Use the LongTerm equivalent instead}

\item{useCovariateConditionOccurrenceInpt180d}{DEPRECATED. Use the ShortTerm equivalent instead}

\item{useCovariateConditionOccurrence30d}{DEPRECATED. Use the ShortTerm equivalent instead}

\item{useCovariateDrugExposure30d}{DEPRECATED. Use the ShortTerm equivalent instead}

\item{useCovariateDrugEra30d}{DEPRECATED. Use the ShortTerm equivalent instead}

\item{useCovariateMeasurement30d}{DEPRECATED. Use the ShortTerm equivalent instead}

\item{useCovariateObservation30d}{DEPRECATED. Use the ShortTerm equivalent instead}

\item{useCovariateProcedureOccurrence30d}{DEPRECATED. Use the ShortTerm equivalent instead}
}
\value{
An object of type \code{defaultCovariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}
