% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDefaultCovariates.R
\name{createCovariateSettings}
\alias{createCovariateSettings}
\title{Create covariate settings}
\usage{
createCovariateSettings(useDemographicsGender = FALSE,
  useDemographicsAge = FALSE, useDemographicsIndexYear = FALSE,
  useDemographicsIndexMonth = FALSE, useConditionOccurrenceLongTerm = FALSE,
  useConditionOccurrenceShortTerm = FALSE, useConditionEraLongTerm = FALSE,
  useConditionEraShortTerm = FALSE, useConditionGroupEraLongTerm = FALSE,
  useConditionGroupEraShortTerm = FALSE, useDrugExposureLongTerm = FALSE,
  useDrugExposureShortTerm = FALSE, useDrugEraLongTerm = FALSE,
  useDrugEraShortTerm = FALSE, useDrugGroupEraLongTerm = FALSE,
  useDrugGroupEraShortTerm = FALSE, useProcedureOccurrenceLongTerm = FALSE,
  useProcedureOccurrenceShortTerm = FALSE,
  useDeviceExposureLongTerm = FALSE, useDeviceExposureShortTerm = FALSE,
  useMeasurementLongTerm = FALSE, useMeasurementShortTerm = FALSE,
  useObservationLongTerm = FALSE, useObservationShortTerm = FALSE,
  useCharlsonIndex = FALSE, longTermDays = 365, shortTermDays = 30,
  windowEndDays = 0, excludedCovariateConceptIds = c(),
  addDescendantsToExclude = TRUE, includedCovariateConceptIds = c(),
  addDescendantsToInclude = TRUE, includedCovariateIds = c(),
  deleteCovariatesSmallCount = 100)
}
\arguments{
\item{longTermDays}{What is the length (in days) of the long-term window?}

\item{shortTermDays}{What is the length (in days) of the short-term window?}

\item{windowEndDays}{What is the last day of the window? 0 means the cohort
start date is the last date (included), 1 means the window
stops the day before the cohort start date, etc.}

\item{useCovariateDemographicsGender}{A boolean value (TRUE/FALSE) to determine if gender
should be included in the model.}

\item{useCovariateDemographicsAge}{A boolean value (TRUE/FALSE) to determine if age
(in 5 year increments) should be included in the
model.}

\item{useCovariateConditionOccurrenceLongTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition in the long term
window prior to or on cohort index date.  Only
applicable if useCovariateConditionOccurrence =
TRUE.}

\item{useCovariateConditionOccurrenceShortTerm}{A boolean value (TRUE/FALSE) to determine if
covariates will be created and used in models that
look for presence/absence of condition in the short term
window prior to or on cohort index date.  Only
applicable if useCovariateConditionOccurrence =
TRUE.}

\item{mediumTermDays}{What is the length (in days) of the medium-term window?}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}
