% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCovariateSettings.R
\name{createCovariateSettings}
\alias{createCovariateSettings}
\title{Create covariate settings}
\usage{
createCovariateSettings(useDemographicsGender = FALSE,
  useDemographicsAge = FALSE, useDemographicsIndexYear = FALSE,
  useDemographicsIndexMonth = FALSE, useConditionOccurrenceLongTerm = FALSE,
  useConditionOccurrenceShortTerm = FALSE, useConditionEraLongTerm = FALSE,
  useConditionEraShortTerm = FALSE, useConditionGroupEraShortTerm = FALSE,
  useConditionGroupEraLongTerm = FALSE, useDrugExposureLongTerm = FALSE,
  useDrugExposureShortTerm = FALSE, useDrugEraLongTerm = FALSE,
  useDrugEraShortTerm = FALSE, useDrugGroupEraLongTerm = FALSE,
  useDrugGroupEraShortTerm = FALSE, useProcedureOccurrenceLongTerm = FALSE,
  useProcedureOccurrenceShortTerm = FALSE,
  useDeviceExposureLongTerm = FALSE, useDeviceExposureShortTerm = FALSE,
  useMeasurementLongTerm = FALSE, useMeasurementShortTerm = FALSE,
  useObservationLongTerm = FALSE, useObservationShortTerm = FALSE,
  useCharlsonIndex = FALSE, longTermStartDays = -365,
  mediumTermStartDays = -180, shortTermStartDays = -30, endDays = 0,
  includedCovariateConceptIds = c(), addDescendantsToInclude = FALSE,
  excludedCovariateConceptIds = c(), addDescendantsToExclude = FALSE,
  includedCovariateIds = c())
}
\arguments{
\item{useDemographicsGender}{Gender of the subject.}

\item{useDemographicsAge}{Age of the subject on the index date (in 5 year
increments).}

\item{useDemographicsIndexYear}{Year of the index date.}

\item{useDemographicsIndexMonth}{Month of the index date.}

\item{useConditionOccurrenceLongTerm}{One covariate per condition in the condition_occurrence
table starting in the long term window.}

\item{useConditionOccurrenceShortTerm}{One covariate per condition in the condition_occurrence
table starting in the short term window.}

\item{useConditionEraLongTerm}{One covariate per condition in the condition_era table
overlapping with any part of the long term window.}

\item{useConditionEraShortTerm}{One covariate per condition in the condition_era table
overlapping with any part of the short term window.}

\item{useConditionGroupEraShortTerm}{One covariate per condition era rolled up to SNOMED groups
in the condition_era table overlapping with any part of the
long term window.}

\item{useConditionGroupEraLongTerm}{One covariate per condition era rolled up to SNOMED groups
in the condition_era table overlapping with any part of the
short term window.}

\item{useDrugExposureLongTerm}{One covariate per drug in the drug_exposure table starting
in the long term window.}

\item{useDrugExposureShortTerm}{One covariate per drug in the drug_exposure table starting
in the short term window.}

\item{useDrugEraLongTerm}{One covariate per drug in the drug_era table overlapping
with any part of the long term window.}

\item{useDrugEraShortTerm}{One covariate per drug in the drug_era table overlapping
with any part of the long short window.}

\item{useDrugGroupEraLongTerm}{One covariate per drug rolled up to ATC groups in the
drug_era table overlapping with any part of the long term
window.}

\item{useDrugGroupEraShortTerm}{One covariate per drug rolled up to ATC groups in the
drug_era table overlapping with any part of the short term
window.}

\item{useProcedureOccurrenceLongTerm}{One covariate per procedure in the procedure_occurrence
table in the long term window.}

\item{useProcedureOccurrenceShortTerm}{One covariate per procedure in the procedure_occurrence
table in the short term window.}

\item{useDeviceExposureLongTerm}{One covariate per device in the device exposure table
starting in the long term window.}

\item{useDeviceExposureShortTerm}{One covariate per device in the device exposure table
starting in the short term window.}

\item{useMeasurementLongTerm}{One covariate per measurement in the measurement table in
the long term window.}

\item{useMeasurementShortTerm}{One covariate per measurement in the measurement table in
the short term window.}

\item{useObservationLongTerm}{One covariate per observation in the observation table in
the long term window.}

\item{useObservationShortTerm}{One covariate per observation in the observation table in
the short term window.}

\item{useCharlsonIndex}{The Charlson comorbidity index (Romano adaptation) using
all conditions prior to the window end.}

\item{longTermStartDays}{What is the start day (relative to the index date) of the
long-term window?}

\item{mediumTermStartDays}{What is the start day (relative to the index date) of the
medium-term window?}

\item{shortTermStartDays}{What is the start day (relative to the index date) of the
short-term window?}

\item{endDays}{What is the end day (relative to the index date) of the
window?}

\item{includedCovariateConceptIds}{A list of concept IDs that should be used to construct
covariates.}

\item{addDescendantsToInclude}{Should descendant concept IDs be added to the list of
concepts to include?}

\item{excludedCovariateConceptIds}{A list of concept IDs that should NOT be used to construct
covariates.}

\item{addDescendantsToExclude}{Should descendant concept IDs be added to the list of
concepts to exclude?}

\item{includedCovariateIds}{A list of covariate IDs that should be restricted to.}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}
