% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefaultCovariateSettings.R
\name{createCovariateSettings}
\alias{createCovariateSettings}
\title{Create covariate settings}
\usage{
createCovariateSettings(useDemographicsGender = FALSE,
  useDemographicsAge = FALSE, useDemographicsAgeGroup = FALSE,
  useDemographicsRace = FALSE, useDemographicsEthnicity = FALSE,
  useDemographicsIndexYear = FALSE, useDemographicsIndexMonth = FALSE,
  useDemographicsPriorObservationTime = FALSE,
  useConditionOccurrenceLongTerm = FALSE,
  useConditionOccurrenceMediumTerm = FALSE,
  useConditionOccurrenceShortTerm = FALSE,
  useConditionOccurrenceInpatientLongTerm = FALSE,
  useConditionOccurrenceInpatientMediumTerm = FALSE,
  useConditionOccurrenceInpatientShortTerm = FALSE,
  useConditionEraLongTerm = FALSE, useConditionEraMediumTerm = FALSE,
  useConditionEraShortTerm = FALSE, useConditionEraOverlapping = FALSE,
  useConditionEraStartLongTerm = FALSE,
  useConditionEraStartMediumTerm = FALSE,
  useConditionEraStartShortTerm = FALSE,
  useConditionSnomedGroupEraLongTerm = FALSE,
  useConditionSnomedGroupEraMediumTerm = FALSE,
  useConditionSnomedGroupEraShortTerm = FALSE,
  useConditionSnomedGroupEraOverlapping = FALSE,
  useConditionSnomedGroupEraStartLongTerm = FALSE,
  useConditionSnomedGroupEraStartMediumTerm = FALSE,
  useConditionSnomedGroupEraStartShortTerm = FALSE,
  useConditionMeddraGroupEraLongTerm = FALSE,
  useConditionMeddraGroupEraMediumTerm = FALSE,
  useConditionMeddraGroupEraShortTerm = FALSE,
  useConditionMeddraGroupEraOverlapping = FALSE,
  useConditionMeddraGroupEraStartLongTerm = FALSE,
  useConditionMeddraGroupEraStartMediumTerm = FALSE,
  useConditionMeddraGroupEraStartShortTerm = FALSE,
  useDrugExposureLongTerm = FALSE, useDrugExposureMediumTerm = FALSE,
  useDrugExposureShortTerm = FALSE, useDrugEraLongTerm = FALSE,
  useDrugEraMediumTerm = FALSE, useDrugEraShortTerm = FALSE,
  useDrugEraOverlapping = FALSE, useDrugEraStartLongTerm = FALSE,
  useDrugEraStartMediumTerm = FALSE, useDrugEraStartShortTerm = FALSE,
  useDrugGroupEraLongTerm = FALSE, useDrugGroupEraMediumTerm = FALSE,
  useDrugGroupEraShortTerm = FALSE, useDrugGroupEraOverlapping = FALSE,
  useDrugGroupEraStartLongTerm = FALSE,
  useDrugGroupEraStartMediumTerm = FALSE,
  useDrugGroupEraStartShortTerm = FALSE,
  useProcedureOccurrenceLongTerm = FALSE,
  useProcedureOccurrenceMediumTerm = FALSE,
  useProcedureOccurrenceShortTerm = FALSE,
  useDeviceExposureLongTerm = FALSE, useDeviceExposureMediumTerm = FALSE,
  useDeviceExposureShortTerm = FALSE, useMeasurementLongTerm = FALSE,
  useMeasurementMediumTerm = FALSE, useMeasurementShortTerm = FALSE,
  useMeasurementValueLongTerm = FALSE,
  useMeasurementValueMediumTerm = FALSE,
  useMeasurementValueShortTerm = FALSE,
  useMeasurementRangeGroupLongTerm = FALSE,
  useMeasurementRangeGroupMediumTerm = FALSE,
  useMeasurementRangeGroupShortTerm = FALSE, useObservationLongTerm = FALSE,
  useObservationMediumTerm = FALSE, useObservationShortTerm = FALSE,
  useCharlsonIndex = FALSE, useDcsi = FALSE, useChads2 = FALSE,
  useChads2Vasc = FALSE, useDistinctConditionCountLongTerm = FALSE,
  useDistinctConditionCountMediumTerm = FALSE,
  useDistinctConditionCountShortTerm = FALSE,
  useDistinctIngredientCountLongTerm = FALSE,
  useDistinctIngredientCountMediumTerm = FALSE,
  useDistinctIngredientCountShortTerm = FALSE,
  useDistinctProcedureCountLongTerm = FALSE,
  useDistinctProcedureCountMediumTerm = FALSE,
  useDistinctProcedureCountShortTerm = FALSE,
  useDistinctMeasurementCountLongTerm = FALSE,
  useDistinctMeasurementCountMediumTerm = FALSE,
  useDistinctMeasurementCountShortTerm = FALSE,
  useVisitCountLongTerm = FALSE, useVisitCountMediumTerm = FALSE,
  useVisitCountShortTerm = FALSE, longTermStartDays = -365,
  mediumTermStartDays = -180, shortTermStartDays = -30, endDays = 0,
  includedCovariateConceptIds = c(), addDescendantsToInclude = FALSE,
  excludedCovariateConceptIds = c(), addDescendantsToExclude = FALSE,
  includedCovariateIds = c())
}
\arguments{
\item{useDemographicsGender}{Gender of the subject.}

\item{useDemographicsAge}{Age of the subject on the index date (in years).}

\item{useDemographicsAgeGroup}{Age of the subject on the index date (in 5 year
age groups)}

\item{useDemographicsRace}{Race of the subject.}

\item{useDemographicsEthnicity}{Ethnicity of the subject.}

\item{useDemographicsIndexYear}{Year of the index date.}

\item{useDemographicsIndexMonth}{Month of the index date.}

\item{useDemographicsPriorObservationTime}{Number of days of observation time preceding the
index date.}

\item{useConditionOccurrenceLongTerm}{One covariate per condition in the
condition_occurrence table starting in the long
term window.}

\item{useConditionOccurrenceMediumTerm}{One covariate per condition in the
condition_occurrence table starting in the medium
term window.}

\item{useConditionOccurrenceShortTerm}{One covariate per condition in the
condition_occurrence table starting in the short
term window.}

\item{useConditionOccurrenceInpatientLongTerm}{One covariate per condition observed in an
inpatient setting in the condition_occurrence
table starting in the long term window.}

\item{useConditionOccurrenceInpatientMediumTerm}{One covariate per condition observed in an
inpatient setting in the condition_occurrence
table starting in the medium term window.}

\item{useConditionOccurrenceInpatientShortTerm}{One covariate per condition observed in an
inpatient setting in the condition_occurrence
table starting in the short term window.}

\item{useConditionEraLongTerm}{One covariate per condition in the condition_era
table overlapping with any part of the long term
window.}

\item{useConditionEraMediumTerm}{One covariate per condition in the condition_era
table overlapping with any part of the medium
term window.}

\item{useConditionEraShortTerm}{One covariate per condition in the condition_era
table overlapping with any part of the short term
window.}

\item{useConditionEraOverlapping}{One covariate per condition in the condition_era
table overlapping with the end of the risk
window.}

\item{useConditionEraStartLongTerm}{One covariate per condition in the condition_era
table starting in the long term window.}

\item{useConditionEraStartMediumTerm}{One covariate per condition in the condition_era
table starting in the medium term window.}

\item{useConditionEraStartShortTerm}{One covariate per condition in the condition_era
table starting in the short term window.}

\item{useConditionSnomedGroupEraLongTerm}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table
overlapping with any part of the long term
window.}

\item{useConditionSnomedGroupEraMediumTerm}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table
overlapping with any part of the medium term
window.}

\item{useConditionSnomedGroupEraShortTerm}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table
overlapping with any part of the short term
window.}

\item{useConditionSnomedGroupEraOverlapping}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table
overlapping with the end of the risk window.}

\item{useConditionSnomedGroupEraStartLongTerm}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table starting
in the long term window.}

\item{useConditionSnomedGroupEraStartMediumTerm}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table starting
in the medium term window.}

\item{useConditionSnomedGroupEraStartShortTerm}{One covariate per condition era rolled up to
SNOMED groups in the condition_era table starting
in the short term window.}

\item{useConditionMeddraGroupEraLongTerm}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table
overlapping with any part of the long term
window.}

\item{useConditionMeddraGroupEraMediumTerm}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table
overlapping with any part of the medium term
window.}

\item{useConditionMeddraGroupEraShortTerm}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table
overlapping with any part of the short term
window.}

\item{useConditionMeddraGroupEraOverlapping}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table
overlapping with the end of the risk window.}

\item{useConditionMeddraGroupEraStartLongTerm}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table starting
in the long term window.}

\item{useConditionMeddraGroupEraStartMediumTerm}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table starting
in the medium term window.}

\item{useConditionMeddraGroupEraStartShortTerm}{One covariate per condition era rolled up to
MedDRA groups in the condition_era table starting
in the short term window.}

\item{useDrugExposureLongTerm}{One covariate per drug in the drug_exposure table
starting in the long term window.}

\item{useDrugExposureMediumTerm}{One covariate per drug in the drug_exposure table
starting in the medium term window.}

\item{useDrugExposureShortTerm}{One covariate per drug in the drug_exposure table
starting in the short term window.}

\item{useDrugEraLongTerm}{One covariate per drug in the drug_era table
overlapping with any part of the long term
window.}

\item{useDrugEraMediumTerm}{One covariate per drug in the drug_era table
overlapping with any part of the medium term
window.}

\item{useDrugEraShortTerm}{One covariate per drug in the drug_era table
overlapping with any part of the short window.}

\item{useDrugEraOverlapping}{One covariate per drug in the drug_era table
overlapping with the end of the risk window.}

\item{useDrugEraStartLongTerm}{One covariate per drug in the drug_era table
starting in the long term window.}

\item{useDrugEraStartMediumTerm}{One covariate per drug in the drug_era table
starting in the medium term window.}

\item{useDrugEraStartShortTerm}{One covariate per drug in the drug_era table
starting in the long short window.}

\item{useDrugGroupEraLongTerm}{One covariate per drug rolled up to ATC groups in
the drug_era table overlapping with any part of
the long term window.}

\item{useDrugGroupEraMediumTerm}{One covariate per drug rolled up to ATC groups in
the drug_era table overlapping with any part of
the medium term window.}

\item{useDrugGroupEraShortTerm}{One covariate per drug rolled up to ATC groups in
the drug_era table overlapping with any part of
the short term window.}

\item{useDrugGroupEraOverlapping}{One covariate per drug rolled up to ATC groups in
the drug_era table overlapping with the end of
the risk window.}

\item{useDrugGroupEraStartLongTerm}{One covariate per drug rolled up to ATC groups in
the drug_era table starting in the long term
window.}

\item{useDrugGroupEraStartMediumTerm}{One covariate per drug rolled up to ATC groups in
the drug_era table starting in the medium term
window.}

\item{useDrugGroupEraStartShortTerm}{One covariate per drug rolled up to ATC groups in
the drug_era table starting in the short term
window.}

\item{useProcedureOccurrenceLongTerm}{One covariate per procedure in the
procedure_occurrence table in the long term
window.}

\item{useProcedureOccurrenceMediumTerm}{One covariate per procedure in the
procedure_occurrence table in the medium term
window.}

\item{useProcedureOccurrenceShortTerm}{One covariate per procedure in the
procedure_occurrence table in the short term
window.}

\item{useDeviceExposureLongTerm}{One covariate per device in the device exposure
table starting in the long term window.}

\item{useDeviceExposureMediumTerm}{One covariate per device in the device exposure
table starting in the medium term window.}

\item{useDeviceExposureShortTerm}{One covariate per device in the device exposure
table starting in the short term window.}

\item{useMeasurementLongTerm}{One covariate per measurement in the measurement
table in the long term window.}

\item{useMeasurementMediumTerm}{One covariate per measurement in the measurement
table in the medium term window.}

\item{useMeasurementShortTerm}{One covariate per measurement in the measurement
table in the short term window.}

\item{useMeasurementValueLongTerm}{One covariate containing the value per
measurement-unit combination in the long term
window.}

\item{useMeasurementValueMediumTerm}{One covariate containing the value per
measurement-unit combination in the medium term
window.}

\item{useMeasurementValueShortTerm}{One covariate containing the value per
measurement-unit combination in the short term
window.}

\item{useMeasurementRangeGroupLongTerm}{Covariates indicating whether measurements are
below, within, or above normal range in the long
term window.}

\item{useMeasurementRangeGroupMediumTerm}{Covariates indicating whether measurements are
below, within, or above normal range in the
medium term window.}

\item{useMeasurementRangeGroupShortTerm}{Covariates indicating whether measurements are
below, within, or above normal range in the short
term window.}

\item{useObservationLongTerm}{One covariate per observation in the observation
table in the long term window.}

\item{useObservationMediumTerm}{One covariate per observation in the observation
table in the medium term window.}

\item{useObservationShortTerm}{One covariate per observation in the observation
table in the short term window.}

\item{useCharlsonIndex}{The Charlson comorbidity index (Romano
adaptation) using all conditions prior to the
window end.}

\item{useDcsi}{The Diabetes Comorbidity Severity Index (DCSI)
using all conditions prior to the window end.}

\item{useChads2}{The CHADS2 score using all conditions prior to
the window end.}

\item{useChads2Vasc}{The CHADS2VASc score using all conditions prior
to the window end.}

\item{useDistinctConditionCountLongTerm}{The number of distinct condition concepts
observed in the long term window.}

\item{useDistinctConditionCountMediumTerm}{The number of distinct condition concepts
observed in the medium term window.}

\item{useDistinctConditionCountShortTerm}{The number of distinct condition concepts
observed in the medium term window.}

\item{useDistinctIngredientCountLongTerm}{The number of distinct ingredients observed in
the long term window.}

\item{useDistinctIngredientCountMediumTerm}{The number of distinct ingredients observed in
the medium term window.}

\item{useDistinctIngredientCountShortTerm}{The number of distinct ingredients observed in
the medium term window.}

\item{useDistinctProcedureCountLongTerm}{The number of distinct procedures observed in the
long term window.}

\item{useDistinctProcedureCountMediumTerm}{The number of distinct procedures observed in the
medium term window.}

\item{useDistinctProcedureCountShortTerm}{The number of distinct procedures observed in the
medium term window.}

\item{useDistinctMeasurementCountLongTerm}{The number of distinct measurements observed in
the long term window.}

\item{useDistinctMeasurementCountMediumTerm}{The number of distinct measurements observed in
the medium term window.}

\item{useDistinctMeasurementCountShortTerm}{The number of distinct measurements observed in
the medium term window.}

\item{useVisitCountLongTerm}{The number of visits observed in the long term
window.}

\item{useVisitCountMediumTerm}{The number of visits observed in the medium term
window.}

\item{useVisitCountShortTerm}{The number of visits observed in the medium term
window.}

\item{longTermStartDays}{What is the start day (relative to the index
date) of the long-term window?}

\item{mediumTermStartDays}{What is the start day (relative to the index
date) of the medium-term window?}

\item{shortTermStartDays}{What is the start day (relative to the index
date) of the short-term window?}

\item{endDays}{What is the end day (relative to the index date)
of the window?}

\item{includedCovariateConceptIds}{A list of concept IDs that should be used to
construct covariates.}

\item{addDescendantsToInclude}{Should descendant concept IDs be added to the
list of concepts to include?}

\item{excludedCovariateConceptIds}{A list of concept IDs that should NOT be used to
construct covariates.}

\item{addDescendantsToExclude}{Should descendant concept IDs be added to the
list of concepts to exclude?}

\item{includedCovariateIds}{A list of covariate IDs that should be restricted
to.}
}
\value{
An object of type \code{covariateSettings}, to be used in other functions.
}
\description{
Create covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}
\examples{
settings <- createCovariateSettings(useDemographicsGender = TRUE,
                                    useDemographicsAge = FALSE,
                                    useDemographicsAgeGroup = TRUE,
                                    useDemographicsRace = TRUE,
                                    useDemographicsEthnicity = TRUE,
                                    useDemographicsIndexYear = TRUE,
                                    useDemographicsIndexMonth = TRUE,
                                    useDemographicsPriorObservationTime = FALSE,
                                    useConditionOccurrenceLongTerm = TRUE,
                                    useConditionOccurrenceMediumTerm = FALSE,
                                    useConditionOccurrenceShortTerm = TRUE,
                                    useConditionOccurrenceInpatientLongTerm = FALSE,
                                    useConditionOccurrenceInpatientMediumTerm = FALSE,
                                    useConditionOccurrenceInpatientShortTerm = FALSE,
                                    useConditionEraLongTerm = TRUE,
                                    useConditionEraMediumTerm = FALSE,
                                    useConditionEraShortTerm = TRUE,
                                    useConditionEraOverlapping = FALSE,
                                    useConditionEraStartLongTerm = FALSE,
                                    useConditionEraStartMediumTerm = FALSE,
                                    useConditionEraStartShortTerm = FALSE,
                                    useConditionSnomedGroupEraLongTerm = TRUE,
                                    useConditionSnomedGroupEraMediumTerm = FALSE,
                                    useConditionSnomedGroupEraShortTerm = TRUE,
                                    useConditionSnomedGroupEraOverlapping = FALSE,
                                    useConditionSnomedGroupEraStartLongTerm = TRUE,
                                    useConditionSnomedGroupEraStartMediumTerm = FALSE,
                                    useConditionSnomedGroupEraStartShortTerm = TRUE,
                                    useConditionMeddraGroupEraLongTerm = TRUE,
                                    useConditionMeddraGroupEraMediumTerm = FALSE,
                                    useConditionMeddraGroupEraShortTerm = TRUE,
                                    useConditionMeddraGroupEraOverlapping = FALSE,
                                    useConditionMeddraGroupEraStartLongTerm = TRUE,
                                    useConditionMeddraGroupEraStartMediumTerm = FALSE,
                                    useConditionMeddraGroupEraStartShortTerm = TRUE,
                                    useDrugExposureLongTerm = FALSE,
                                    useDrugExposureMediumTerm = FALSE,
                                    useDrugExposureShortTerm = FALSE,
                                    useDrugEraLongTerm = TRUE,
                                    useDrugEraMediumTerm = FALSE,
                                    useDrugEraShortTerm = TRUE,
                                    useDrugEraOverlapping = TRUE,
                                    useDrugEraStartLongTerm = FALSE,
                                    useDrugEraStartMediumTerm = FALSE,
                                    useDrugEraStartShortTerm = FALSE,
                                    useDrugGroupEraLongTerm = TRUE,
                                    useDrugGroupEraMediumTerm = FALSE,
                                    useDrugGroupEraShortTerm = TRUE,
                                    useDrugGroupEraOverlapping = TRUE,
                                    useDrugGroupEraStartLongTerm = FALSE,
                                    useDrugGroupEraStartMediumTerm = FALSE,
                                    useDrugGroupEraStartShortTerm = FALSE,
                                    useProcedureOccurrenceLongTerm = TRUE,
                                    useProcedureOccurrenceMediumTerm = FALSE,
                                    useProcedureOccurrenceShortTerm = TRUE,
                                    useDeviceExposureLongTerm = TRUE,
                                    useDeviceExposureMediumTerm = FALSE,
                                    useDeviceExposureShortTerm = TRUE,
                                    useMeasurementLongTerm = TRUE,
                                    useMeasurementMediumTerm = FALSE,
                                    useMeasurementShortTerm = TRUE,
                                    useMeasurementValueLongTerm = FALSE,
                                    useMeasurementValueMediumTerm = FALSE,
                                    useMeasurementValueShortTerm = FALSE,
                                    useMeasurementRangeGroupLongTerm = TRUE,
                                    useMeasurementRangeGroupMediumTerm = FALSE,
                                    useMeasurementRangeGroupShortTerm = FALSE,
                                    useObservationLongTerm = TRUE,
                                    useObservationMediumTerm = FALSE,
                                    useObservationShortTerm = TRUE,
                                    useCharlsonIndex = TRUE,
                                    useDcsi = TRUE,
                                    useChads2 = TRUE,
                                    useChads2Vasc = TRUE,
                                    useDistinctConditionCountLongTerm = FALSE,
                                    useDistinctConditionCountMediumTerm = FALSE,
                                    useDistinctConditionCountShortTerm = FALSE,
                                    useDistinctIngredientCountLongTerm = FALSE,
                                    useDistinctIngredientCountMediumTerm = FALSE,
                                    useDistinctIngredientCountShortTerm = FALSE,
                                    useDistinctProcedureCountLongTerm = FALSE,
                                    useDistinctProcedureCountMediumTerm = FALSE,
                                    useDistinctProcedureCountShortTerm = FALSE,
                                    useDistinctMeasurementCountLongTerm = FALSE,
                                    useDistinctMeasurementCountMediumTerm = FALSE,
                                    useDistinctMeasurementCountShortTerm = FALSE,
                                    useVisitCountLongTerm = FALSE,
                                    useVisitCountMediumTerm = FALSE,
                                    useVisitCountShortTerm = FALSE,
                                    longTermStartDays = -365,
                                    mediumTermStartDays = -180,
                                    shortTermStartDays = -30,
                                    endDays = 0,
                                    includedCovariateConceptIds = c(),
                                    addDescendantsToInclude = FALSE,
                                    excludedCovariateConceptIds = c(),
                                    addDescendantsToExclude = FALSE,
                                    includedCovariateIds = c())

}
