% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateData.R
\name{createEmptyCovariateData}
\alias{createEmptyCovariateData}
\title{Create an empty CovariateData object}
\usage{
createEmptyCovariateData(cohortIds, aggregated, temporal)
}
\arguments{
\item{cohortIds}{For which cohort IDs should the covariate data be created?}

\item{aggregated}{if the covariate data is aggregated}

\item{temporal}{if the covariate data is temporal}
}
\value{
the empty CovariateData object
}
\description{
Create an empty CovariateData object
}
