% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTemporalCovariates.R
\name{createTemporalCovariateSettings}
\alias{createTemporalCovariateSettings}
\title{Create temporal covariate settings}
\usage{
createTemporalCovariateSettings(useCovariateConditionEraStart = FALSE,
  useCovariateConditionEraPresent = FALSE,
  excludedCovariateConceptIds = c(), includedCovariateConceptIds = c(),
  startDays = -365:-1, endDays = -364:0)
}
\arguments{
\item{excludedCovariateConceptIds}{A list of concept IDs that should NOT be used to
construct covariates.}

\item{includedCovariateConceptIds}{A list of concept IDs that should be used to
construct covariates.}
}
\value{
An object of type \code{defaultCovariateSettings}, to be used in other functions.
}
\description{
Create temporal covariate settings
}
\details{
creates an object specifying how covariates should be contructed from data in the CDM model.
}

