% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureExtraction.R
\docType{package}
\name{FeatureExtraction-package}
\alias{FeatureExtraction}
\alias{FeatureExtraction-package}
\title{FeatureExtraction: Generating Features for a Cohort}
\description{
An R package for generating features (covariates) for a cohort using data in the Common Data Model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/FeatureExtraction}
  \item Report bugs at \url{https://github.com/OHDSI/FeatureExtraction/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Marc Suchard
  \item Patrick Ryan
  \item Jenna Reps
}

}
\keyword{internal}
