% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterByCohortDefinitionId}
\alias{filterByCohortDefinitionId}
\title{Filter covariates by cohort definition ID}
\usage{
filterByCohortDefinitionId(covariateData, cohortId)
}
\arguments{
\item{covariateData}{An object of type \code{CovariateData}}

\item{cohortId}{The cohort definition ID to keep.}
}
\value{
An object of type \code{covariateData}.
}
\description{
Filter covariates by cohort definition ID
}
\examples{
\dontrun{
covariateData <- FeatureExtraction:::createEmptyCovariateData(
  cohortId = 1,
  aggregated = FALSE,
  temporal = FALSE
)

covData <- filterByCohortDefinitionId(
  covariateData = covariateData,
  cohortId = 1
)
}

}
