% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterByCohortDefinitionId}
\alias{filterByCohortDefinitionId}
\title{Filter covariates by cohort definition IDs}
\usage{
filterByCohortDefinitionId(covariateData, cohortId = 1, cohortIds = c(1))
}
\arguments{
\item{covariateData}{An object of type \code{CovariateData}}

\item{cohortId}{DEPRECATED The cohort definition IDs to keep.}

\item{cohortIds}{The cohort definition IDs to keep.}
}
\value{
An object of type \code{covariateData}.
}
\description{
Filter covariates by cohort definition IDs
}
\examples{
\donttest{
covariateData <- FeatureExtraction::createEmptyCovariateData(
  cohortIds = c(1, 2),
  aggregated = TRUE,
  temporal = FALSE
)

covData <- filterByCohortDefinitionId(
  covariateData = covariateData,
  cohortIds = c(1)
)
}

}
