% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterByCohortDefinitionIds}
\alias{filterByCohortDefinitionIds}
\title{Filter covariates by cohort definition IDs}
\usage{
filterByCohortDefinitionIds(covariateData, cohortIds)
}
\arguments{
\item{covariateData}{An object of type \code{CovariateData}}

\item{cohortIds}{The cohort definitions IDs to keep.}
}
\value{
An object of type \code{covariateData}.
}
\description{
Filter covariates by cohort definition IDs
}
