% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{filterByRowId}
\alias{filterByRowId}
\title{Filter covariates by row ID}
\usage{
filterByRowId(covariateData, rowIds)
}
\arguments{
\item{covariateData}{An object of type \code{CovariateData}}

\item{rowIds}{A vector containing the rowIds to keep.}
}
\value{
An object of type \code{covariateData}.
}
\description{
Filter covariates by row ID
}
