% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovariateData.R
\name{isTemporalCovariateData}
\alias{isTemporalCovariateData}
\title{Check whether covariate data is temporal}
\usage{
isTemporalCovariateData(x)
}
\arguments{
\item{x}{The covariate data object to check.}
}
\value{
A logical value.
}
\description{
Check whether covariate data is temporal
}
