% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{tidyCovariateData}
\alias{tidyCovariateData}
\title{Tidy covariate data}
\usage{
tidyCovariateData(covariateData, covariates, covariateRef, populationSize,
  minFraction = 0, normalize = TRUE, removeRedundancy = TRUE)
}
\arguments{
\item{covariateData}{An object as generated using the \code{\link{getDbCovariateData}}
function. If provided, the \code{covariates}, \code{covariateRef}, and 
\code{populationSize} arguments will be ignored.}

\item{covariates}{An ffdf object with the covariate values in spare format. Will be 
ignored if \code{covariateData} is provided.}

\item{covariateRef}{An ffdf object with the covariate definitions. Will be 
ignored if \code{covariateData} is provided. Only needed when 
\code{removeRedundancy = TRUE}.}

\item{populationSize}{An integer specifying the total number of unique cohort entries (rowIds). Will be 
ignored if \code{covariateData} is provided. Only needed when 
\code{removeRedundancy = TRUE}.}

\item{minFraction}{Minimum fraction of the population that should have a non-zero value for a covariate
for that covariate to be kept. Set to 0 to don't filter on frequency.}

\item{normalize}{Normalize the coviariates? (dividing by the max)}

\item{removeRedundancy}{Should redundant covariates be removed?}
}
\description{
Tidy covariate data
}
\details{
Normalize covariate values by dividing by the max and/or remove redundant covariates.
}
